Try a more conservative strategy for eta-reduction
Here I /do not/ eta-reduce to PAPs, becuase that
has some down-sides; notably that
  \x. f y x  ==>  f y
might mean that 'f' is no longer applied to two
args, and won't inline any more.  (Imagine that
y is let-bound to a value.)

By drawing back on this I want to see the effect of
everything else.